Remove `getSequenceFiles` helper file; make it a method of `SequenceFactory`
