+++ packages/11ty/_plugins/figures/sequence/factory.js
@@ -1,4 +1,5 @@
-const getSequenceFiles = require('./get-sequence-files')
+const fs = require('fs')
+const path = require('path')
 const Sequence = require('./index')
 
 /**
@@ -27,5 +28,27 @@ module.exports = class SequenceFactory {
       }
       return new Sequence(this.figure, sequence, files)
     })
+  }
+
+  /**
+   * Read the directory where sequence images exist to create an array of filenames
+   *
+   * @param  {Object} figure     Figure data
+   * @param  {Object} iiifConfig IIIF Config data
+   * @return {Array<string>}     An array of filenames
+   */
+  getSequenceFiles(sequence, iiifConfig) {
+    if (!sequence) return
+    const { dirs } = iiifConfig
+    const { imagesDir, inputRoot } = dirs
+    const sequenceDir = path.join(inputRoot, imagesDir, sequence.id)
+    const defaultSequenceRegex = /^\d{3}\.(jpg|png)$/
+    const sequenceRegex = sequence.regex
+      ? new RegExp(sequence.regex.slice(1, -1))
+      : defaultSequenceRegex
+    if (!fs.existsSync(sequenceDir)) return
+    return fs
+      .readdirSync(sequenceDir)
+      .filter((sequenceItemFilename) => sequenceItemFilename.match(sequenceRegex))
   }
 }
